backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
httpx==0.28.1
paramiko==3.5.0
pydantic==2.10.3
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
cryptography==44.0.0
python-multipart==0.0.20
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
